Fix eggregious bug with Twitter::Cursor whereby reached_limit? will never return false [Issue sferik/twitter#929]
Twitter::Cursor replaces attrs on every pagination iteration, which
means that attrs[@key].count will *never* be below the max number of
entries per page.

I'm sure the intent here was to have reached_limit? return true once the
total number of entries loaded reached that limit, but that's being
tracked in @collection, not attrs.

This fixes this problem by checking @collection.count for
reached_limit?.

This was manifesting in calls to direct_messages_events overflowing
rate limiting, which is only 15 for this endpoint. If you had more than
15 pages worth of results (750 events), this would thus consistently
fail with a rate limit exception.